chore: move hooks to ui hooks folder
